mcp_client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.4.0",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
